Also use BorrowedFd in SockMapFd

+++ aya/src/programs/trace_point.rs
@@ -85,7 +85,7 @@ impl TracePoint {
                 call: "perf_event_open".to_owned(),
                 io_error,
             }
-        })? as i32;
+        })?;
 
         let link = perf_attach(self.data.fd_or_err()?, fd)?;
         self.data.links.insert(TracePointLink::new(link))
